requirements.txt
streamlit>=1.25,<2.0
sqlalchemy>=1.4,<2.0
python-dotenv>=1.0.0
alembic>=1.10
google-generativeai>=0.7.0
